# Process compact rules lazily