Task/Issue URL:

## Description:
Processing a large number of compact rules during page load may have an effect on performance: local testing saw that going from ~300 to ~1000 rules increased processing time from 3 to 6.5ms.

Given that we only need a small subset of the rule attributes during a single page lifecycle, we can encapsulate the rule in a lazy class, that only parses the value when accessed. This should also have a smaller memory footprint, as we never need to expand the full ruleset.

Testing the same rulesets as before, parse time goes from 0.45 to 0.85ms. The overall improvement for the `initialize` function is 13ms (before this change) to 8ms.